# qrcode/api/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package http

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"io"
	"mime/multipart"
	"net/http"

	common "htqrcode/common"

	"github.com/labstack/echo/v5"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorDetail defines model for ErrorDetail.
type ErrorDetail = common.HttpErrorDetail

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse = common.HttpErrorResponse

// MessageRequest Fields are required according to type. wifi needs wifi_name/password; sms needs phone_number/message; mail and mailto need email/subject/body; geo needs latitude/longitude; tel needs phone_number; contact needs name/phone/email; vcard also needs organization/title/address/website; calendar and ical need summary/start_time/end_time; text needs text.
type MessageRequest struct {
	Address      string `json:"address,omitempty"`
	Body         string `json:"body,omitempty"`
	Description  string `json:"description,omitempty"`
	Email        string `json:"email,omitempty"`
	EndTime      string `json:"end_time,omitempty"`
	Label        string `json:"label,omitempty"`
	Latitude     string `json:"latitude,omitempty"`
	Location     string `json:"location,omitempty"`
	Longitude    string `json:"longitude,omitempty"`
	Message      string `json:"message,omitempty"`
	Name         string `json:"name,omitempty"`
	Organization string `json:"organization,omitempty"`
	Password     string `json:"password,omitempty"`
	Phone        string `json:"phone,omitempty"`
	PhoneNumber  string `json:"phone_number,omitempty"`
	StartTime    string `json:"start_time,omitempty"`
	Subject      string `json:"subject,omitempty"`
	Summary      string `json:"summary,omitempty"`
	Text         string `json:"text,omitempty"`
	Title        string `json:"title,omitempty"`
	Type         string `json:"type,omitempty"`
	Website      string `json:"website,omitempty"`
	WifiName     string `json:"wifi_name,omitempty"`
}

// PostV1GenerateQrCodeMultipartBody defines parameters for PostV1GenerateQrCode.
type PostV1GenerateQrCodeMultipartBody struct {
	// BorderWidth Optional border width (0–1024).
	BorderWidth string `json:"border_width,omitempty"`

	// Data JSON-encoded MessageRequest. Unknown or omitted types use the text field.
	//
	// Example: {"type":"text","text":"https://example.com"}
	Data string `json:"data"`

	// ForegroundColor Optional six-digit RGB hex foreground color, with or without a leading hash.
	ForegroundColor string `json:"foreground_color,omitempty"`

	// HalftoneImg Optional PNG/JPEG halftone (up to 5 MiB and 4096x4096 pixels). Ignored in circle mode.
	HalftoneImg openapi_types.File `json:"halftone_img,omitempty"`

	// IsCircleShape Optional boolean string for circle shape mode.
	IsCircleShape string `json:"is_circle_shape,omitempty"`

	// IsCustomShape Optional boolean string for custom shape mode.
	IsCustomShape string `json:"is_custom_shape,omitempty"`

	// LogoImg Optional PNG/JPEG logo (up to 1 MiB and 500x500 pixels).
	LogoImg openapi_types.File `json:"logo_img,omitempty"`

	// QrWidth Optional QR width (6–255).
	QrWidth string `json:"qr_width,omitempty"`
}

// PostV1GenerateQrCodeMultipartRequestBody defines body for PostV1GenerateQrCode for multipart/form-data ContentType.
type PostV1GenerateQrCodeMultipartRequestBody PostV1GenerateQrCodeMultipartBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// PostV1GenerateQrCode Generate QR code from form data
	// (POST /api/v1/generate-qrcode)
	PostV1GenerateQrCode(ctx *echo.Context) error
	// GetV1VeryFirstApi Sample hello API
	// (GET /api/v1/very-first-api)
	GetV1VeryFirstApi(ctx *echo.Context) error
	// GetHealth Health check
	// (GET /health)
	GetHealth(ctx *echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostV1GenerateQrCode converts echo context to params.
func (w *ServerInterfaceWrapper) PostV1GenerateQrCode(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostV1GenerateQrCode(ctx)
	return err
}

// GetV1VeryFirstApi converts echo context to params.
func (w *ServerInterfaceWrapper) GetV1VeryFirstApi(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetV1VeryFirstApi(ctx)
	return err
}

// GetHealth converts echo context to params.
func (w *ServerInterfaceWrapper) GetHealth(ctx *echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetHealth(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/health", wrapper.GetHealth, options.OperationMiddlewares["getHealth"]...)
	router.GET(options.BaseURL+"/api/v1/very-first-api", wrapper.GetV1VeryFirstApi, options.OperationMiddlewares["getV1VeryFirstApi"]...)
	router.POST(options.BaseURL+"/api/v1/generate-qrcode", wrapper.PostV1GenerateQrCode, options.OperationMiddlewares["postV1GenerateQrCode"]...)

}

type PostV1GenerateQrCodeRequestObject struct {
	Body *multipart.Reader
}

type PostV1GenerateQrCodeResponseObject interface {
	VisitPostV1GenerateQrCodeResponse(w http.ResponseWriter) error
}

type PostV1GenerateQrCode200ImagepngResponse struct {
	Body          io.Reader
	ContentLength int64
}

func (response PostV1GenerateQrCode200ImagepngResponse) VisitPostV1GenerateQrCodeResponse(w http.ResponseWriter) error {

	w.Header().Set("Content-Type", "image/png")
	if response.ContentLength != 0 {
		w.Header().Set("Content-Length", fmt.Sprint(response.ContentLength))
	}
	w.WriteHeader(200)

	if closer, ok := response.Body.(io.ReadCloser); ok {
		defer closer.Close()
	}
	_, err := io.Copy(w, response.Body)
	return err
}

type PostV1GenerateQrCode400JSONResponse ErrorResponse

func (response PostV1GenerateQrCode400JSONResponse) VisitPostV1GenerateQrCodeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PostV1GenerateQrCode500JSONResponse ErrorResponse

func (response PostV1GenerateQrCode500JSONResponse) VisitPostV1GenerateQrCodeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type PostV1GenerateQrCodedefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response PostV1GenerateQrCodedefaultJSONResponse) VisitPostV1GenerateQrCodeResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetV1VeryFirstApiRequestObject struct {
}

type GetV1VeryFirstApiResponseObject interface {
	VisitGetV1VeryFirstApiResponse(w http.ResponseWriter) error
}

type GetV1VeryFirstApi200TextResponse string

func (response GetV1VeryFirstApi200TextResponse) VisitGetV1VeryFirstApiResponse(w http.ResponseWriter) error {

	w.Header().Set("Content-Type", "text/plain")
	w.WriteHeader(200)

	_, err := w.Write([]byte(fmt.Sprint(response)))
	return err
}

type GetV1VeryFirstApidefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response GetV1VeryFirstApidefaultJSONResponse) VisitGetV1VeryFirstApiResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

type GetHealthRequestObject struct {
}

type GetHealthResponseObject interface {
	VisitGetHealthResponse(w http.ResponseWriter) error
}

type GetHealth200TextResponse string

func (response GetHealth200TextResponse) VisitGetHealthResponse(w http.ResponseWriter) error {

	w.Header().Set("Content-Type", "text/plain")
	w.WriteHeader(200)

	_, err := w.Write([]byte(fmt.Sprint(response)))
	return err
}

type GetHealthdefaultJSONResponse struct {
	Body       ErrorResponse
	StatusCode int
}

func (response GetHealthdefaultJSONResponse) VisitGetHealthResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response.Body); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// PostV1GenerateQrCode Generate QR code from form data
	// (POST /api/v1/generate-qrcode)
	PostV1GenerateQrCode(ctx context.Context, request PostV1GenerateQrCodeRequestObject) (PostV1GenerateQrCodeResponseObject, error)
	// GetV1VeryFirstApi Sample hello API
	// (GET /api/v1/very-first-api)
	GetV1VeryFirstApi(ctx context.Context, request GetV1VeryFirstApiRequestObject) (GetV1VeryFirstApiResponseObject, error)
	// GetHealth Health check
	// (GET /health)
	GetHealth(ctx context.Context, request GetHealthRequestObject) (GetHealthResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PostV1GenerateQrCode operation middleware
func (sh *strictHandler) PostV1GenerateQrCode(ctx *echo.Context) error {
	var request PostV1GenerateQrCodeRequestObject

	if reader, err := ctx.Request().MultipartReader(); err != nil {
		return err
	} else {
		request.Body = reader
	}

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostV1GenerateQrCode(ctx.Request().Context(), request.(PostV1GenerateQrCodeRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostV1GenerateQrCode")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostV1GenerateQrCodeResponseObject); ok {
		return validResponse.VisitPostV1GenerateQrCodeResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetV1VeryFirstApi operation middleware
func (sh *strictHandler) GetV1VeryFirstApi(ctx *echo.Context) error {
	var request GetV1VeryFirstApiRequestObject

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetV1VeryFirstApi(ctx.Request().Context(), request.(GetV1VeryFirstApiRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetV1VeryFirstApi")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetV1VeryFirstApiResponseObject); ok {
		return validResponse.VisitGetV1VeryFirstApiResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// GetHealth operation middleware
func (sh *strictHandler) GetHealth(ctx *echo.Context) error {
	var request GetHealthRequestObject

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.GetHealth(ctx.Request().Context(), request.(GetHealthRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetHealth")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(GetHealthResponseObject); ok {
		return validResponse.VisitGetHealthResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}